requirements.txt
absl-py==0.12.0
anyio @ file:///tmp/build/80754af9/anyio_1617783275907/work/dist
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1613037097816/work
astunparse==1.6.3
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
Babel @ file:///tmp/build/80754af9/babel_1607110387436/work
backcall==0.2.0
bleach==3.1.5
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1613246945912/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
cryptography @ file:///tmp/build/80754af9/cryptography_1616769286105/work
cycler==0.10.0
decorator==4.4.2
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints==0.3
flatbuffers==1.12
gast==0.3.3
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imageio==2.16.1
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1617874469820/work
ipykernel==5.3.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
joblib==1.1.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyter-packaging @ file:///tmp/build/80754af9/jupyter-packaging_1613502826984/work
jupyter-server @ file:///tmp/build/80754af9/jupyter_server_1616083640759/work
jupyterlab @ file:///tmp/build/80754af9/jupyterlab_1617303052588/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1617134334258/work
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
libclang==13.0.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
nbclassic @ file:///tmp/build/80754af9/nbclassic_1616085367084/work
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert==5.6.1
nbformat==5.0.7
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook==6.0.3
numpy==1.22.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.15.8
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent==0.16.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.3.1
pyzmq==19.0.1
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.23.1
scipy==1.8.0
Send2Trash==1.5.0
six @ file:///tmp/build/80754af9/six_1605205327372/work
sklearn==0.0
sniffio @ file:///tmp/build/80754af9/sniffio_1614030475067/work
tensorboard==2.8.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.8.0
tensorflow-estimator==2.4.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions==3.7.4.3
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
